Move method extensions to CodeGen.Extensions
