+++ src/MagicOnion.GeneratorCore/CodeGen/Extensions/MagicOnionServiceInfoExtensions.cs
@@ -1,0 +1,18 @@
+using MagicOnion.Generator.CodeAnalysis;
+using System.Text.RegularExpressions;
+
+namespace MagicOnion.Generator.CodeGen.Extensions
+{
+    public static class MagicOnionServiceInfoExtensions
+    {
+        // NOTE: A client name is derived from original interface name without 'I' prefix.
+        // - ImportantService  --> ImportantServiceClient
+        // - IImportantService --> ImportantServiceClient
+        // - I0123Service      --> I0123ServiceClient
+        public static string GetClientName(this IMagicOnionServiceInfo serviceInfo)
+            => (Regex.IsMatch(serviceInfo.ServiceType.Name, "I[^a-z0-9]") ? serviceInfo.ServiceType.Name.Substring(1) : serviceInfo.ServiceType.Name) + "Client";
+
+        public static string GetClientFullName(this IMagicOnionServiceInfo serviceInfo)
+            => (serviceInfo.ServiceType.Namespace != null ? serviceInfo.ServiceType.Namespace + "." : "") + serviceInfo.GetClientName();
+    }
+}
