+++ src/MagicOnion.GeneratorCore/CodeGen/RegisterTemplate.cs
@@ -12,7 +12,7 @@ namespace MagicOnion.Generator.CodeGen
     using System.Linq;
     using System.Text;
     using System.Collections.Generic;
-    using MagicOnion.Generator.CodeAnalysis;
+    using MagicOnion.Generator.CodeGen.Extensions;
     using System;
     
     /// <summary>
