+++ src/MagicOnion.GeneratorCore/CodeGen/RegisterTemplate.tt
@@ -3,7 +3,7 @@
 <#@ import namespace="System.Linq" #>
 <#@ import namespace="System.Text" #>
 <#@ import namespace="System.Collections.Generic" #>
-<#@ import namespace="MagicOnion.Generator.CodeAnalysis" #>
+<#@ import namespace="MagicOnion.Generator.CodeGen.Extensions" #>
 #pragma warning disable 618
 #pragma warning disable 612
 #pragma warning disable 414
