+++ src/MagicOnion.GeneratorCore/CodeGen/StaticMagicOnionClientGenerator.cs
@@ -2,8 +2,8 @@ using System.CodeDom.Compiler;
 using System.Collections.Generic;
 using System.IO;
 using System.Linq;
-using System.Text;
 using MagicOnion.Generator.CodeAnalysis;
+using MagicOnion.Generator.CodeGen.Extensions;
 using MagicOnion.Generator.Internal;
 
 namespace MagicOnion.Generator.CodeGen
